organize demo ai app vertically
