+++ template/app/src/demo-ai-app/schedule.ts
