--- template/app/src/server/queries.ts
@@ -1,32 +1,0 @@
-import { type GptResponse, type Task } from 'wasp/entities';
-import { HttpError } from 'wasp/server';
-import { type GetGptResponses, type GetAllTasksByUser } from 'wasp/server/operations';
-
-export const getGptResponses: GetGptResponses<void, GptResponse[]> = async (args, context) => {
-  if (!context.user) {
-    throw new HttpError(401);
-  }
-  return context.entities.GptResponse.findMany({
-    where: {
-      user: {
-        id: context.user.id,
-      },
-    },
-  });
-};
-
-export const getAllTasksByUser: GetAllTasksByUser<void, Task[]> = async (_args, context) => {
-  if (!context.user) {
-    throw new HttpError(401);
-  }
-  return context.entities.Task.findMany({
-    where: {
-      user: {
-        id: context.user.id,
-      },
-    },
-    orderBy: {
-      createdAt: 'desc',
-    },
-  });
-};
